Crate goji

Goji provides an interface for Jira's REST api

Re-exports

pub use crate::issues::*;

pub use crate::boards::*;

pub use crate::sprints::*;

Modules

boards
Interfaces for accessing and managing boards

issues
Interfaces for accessing and managing issues

resolution
Interfaces for accessing and managing resolutions

sprints
Interfaces for accessing and managing sprints

Structs

Attachment

Changelog

Comment

Comments

EmptyResponse

Errors
represents an general jira error response

History

HistoryItem

Issue
represents a single jira issue

IssueLink
represents link relationship between issues

Jira
Entrypoint into client interface https://docs.atlassian.com/jira/REST/latest/

LinkType
represents type of issue relation

Resolution

Search
Search interface

SearchOptions
options availble for search

SearchOptionsBuilder
a builder interface for search option Typically this is initialized with SearchOptions::builder()

SearchResults

Status

Transition

TransitionOption

TransitionOptions
contains list of options an issue can transitions through

TransitionTo

TransitionTriggerOptions

TransitionTriggerOptionsBuilder

Transitions
issue transition interface

User

Version

Visibility

Enums

Credentials
Types of authentication credentials

Error
an enumeration over potential errors that may happen when sending a request to jira

Type Definitions

Result